requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.3.4
    # via aiohttp
aiohttp==3.10.0
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
attrs==23.2.0
    # via aiohttp
blinker==1.8.2
    # via flask
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
dataclasses-json==0.6.7
    # via langchain-community
environs==9.5.0
    # via pymilvus
filelock==3.15.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flask==3.0.3
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.6.1
    # via
    #   huggingface-hub
    #   torch
greenlet==3.0.3
    # via sqlalchemy
grpcio==1.63.0
    # via pymilvus
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via ollama
huggingface-hub==0.24.5
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via
    #   flask
    #   torch
joblib==1.4.2
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.2.11
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.2.10
    # via -r requirements.in
langchain-core==0.2.26
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.2.2
    # via
    #   -r requirements.in
    #   langchain
langsmith==0.1.95
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.21.3
    # via
    #   dataclasses-json
    #   environs
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
    #   pandas
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
ollama==0.3.1
    # via -r requirements.in
orjson==3.10.6
    # via langsmith
packaging==24.1
    # via
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   transformers
pandas==2.2.2
    # via
    #   -r requirements.in
    #   pymilvus
pillow==10.4.0
    # via sentence-transformers
protobuf==5.27.3
    # via pymilvus
pydantic==2.8.2
    # via
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.20.1
    # via pydantic
pymilvus==2.4.4
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   environs
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   transformers
regex==2024.7.24
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langsmith
    #   transformers
safetensors==0.4.3
    # via transformers
scikit-learn==1.4.0
    # via
    #   -r requirements.in
    #   sentence-transformers
scipy==1.14.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.0.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.31
    # via
    #   langchain
    #   langchain-community
sympy==1.13.1
    # via torch
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.19.1
    # via transformers
torch==2.4.0
    # via sentence-transformers
tqdm==4.66.4
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.43.3
    # via sentence-transformers
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
ujson==5.10.0
    # via pymilvus
urllib3==2.2.2
    # via requests
werkzeug==3.0.3
    # via flask
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
